+++ packages/yijson/CHANGELOG.md
@@ -1,5 +1,17 @@
 # @yicode/yijson-mini
 
+## 1.2.1
+
+### Patch Changes
+
+-   .
+
+## 1.2.0
+
+### Minor Changes
+
+-   .
+
 ## 1.1.2
 
 ### Patch Changes